chore: Add optimization to #4716 (#4718)
# Description

## Problem\*


## Summary\*

Adds the actual array set optimization to
https://github.com/noir-lang/noir/pull/4716. This is separate since the
other PR still has errors and I want to fix those before adding more
with the actual optimization. At the same time, I've put up this PR to
see the circuit size changes in CI.

## Additional Context

This PR is currently failing a few additional tests. I think these are
mostly array out of bounds accesses from storing which indices were
changed. In some if statements, these indices may be out of bounds but
not error since that if case was not run. Trying to get and store to
this same index after the if then gives an OOB error. This can be fixed
by only applying the optimization when constant indices are used and
filtering out the OOB ones but then dynamic indices won't see this
important optimization at all.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.