chore: create placeholder version of 0.26.0 docs (#4782)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR copies the 0.27.0 docs into the 0.26.0 directory so that the
build passes again. I went for this route as for some reason the 0.26.0
commit is not building for me.

cc @noir-lang/developerrelations as there may be changes between these
two versions which won't be reflected in these docs.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.