chore: Rename 'global' to 'function' in the monomorphization pass (#4774)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

The original intention of this was to store functions and globals but
after globals were implemented, only functions ended up being used in
the `globals` hashmap. We also probably don't want to monomorphize
globals into different copies going forward so I renamed the map to
`functions` and clarified references to it to avoid confusion.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.