chore: Use BrilligCall for unconstrained main and update AVM transpiler (#5797)
A follow-up to https://github.com/AztecProtocol/aztec-packages/pull/5737

Separate PR as this touches mainly the AVM transpiler and not Noir
codegen itself. Look in PR #5737 for full details about the switch, but
basically we are moving away from the `Brillig` opcode to a
`BrilligCall` opcode that contains a Brillig call opcode. The AVM needs
to be updated to account for this change.