fix: issue 4682 and add solver for unconstrained bigintegers (#4729)
# Description

## Problem\*

Managing bigintegers as ID makes it difficult to handle them in
conditionals and also when passing them around unconstrained functions.
As a matter of fact these two use cases do not work, the first one being
documented in issue #4682, while the second one was never implemented.

## Summary\*

In order to solve this two issues, I modified the bigint representation
so that it stores its byte array instead of an id, allowing it to work
directly with conditional and unconstrained functions.


## Additional Context

For simplicity, I also made the byte arrays of hardcoded length 32.
Since the moduli we support are of this length and since custom
bigintegers will be implemented differently, I do not expect this length
to change.
The barrentenberg interface has not changed and we still use id for it.
The Id is constructed when required by doing a call to from_bytes.

## Documentation\*
The traits have not changed, I have simply added a from_byte_32 method
which takes a 32-bytes array as input.

Check one:
- [] No documentation needed.
- [X] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>