chore: add benchmarks for serializing a dummy program (#4813)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds some benchmarks for serializing a (not particularly
realistic) ACIR program to give a sense of how this scales with the
number of opcodes in a program. As part of this I've fixed some issues
in the existing benchmarks which prevented good flamegraphs being
generated on linux.

I've also smoothed out the experience for creating flamegraphs with new
scripts

1. run `sudo ./scripts/benchmark_start.sh`
2. cargo bench --bench serialization  -- --profile-time=10
3. run `sudo ./scripts/benchmark_stop.sh`

The flamegraphs for all of the serialization benchmarks will then be
stored in `target/criterion/<benchmark_name>/profile`

The scripts to start and stop benchmarking disable/enable some kernel
settings which improve profiling performance at the expense of security.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.