feat: implement `Eq` trait on `BoundedVec` (#4830)
# Description

## Problem\*

Resolves #4827 

## Summary\*

We're currently not implementing the `Eq` trait on `BoundedVec`, which
is added in this PR.

I've opted for ignoring differences in the storage array past `len` as
while it's not possible to have non-zeroed entries past this index if
you use the proper interface, it's possible to construct `BoundedVec`s
for which this is not the case. Once we allow proper encapsulation of
the internal state then we could simply check that the underlying
storage is equal.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>