Add "negative test" for addition
The point of this and the following tests to be added is the following:
if the tests are only expected to succeed, it might be because the code
is never actually ran. This is because the prove command outputs nothing
when it succeeds.

Adding this test verifies that the assertion in the code does not hold.
As it shouldn't. This is confusing, I know.