feat(experimental): Add `comptime` keyword (#4840)
# Description

## Problem\*

Resolves #4588

## Summary\*

Adds the `comptime` keyword which can be used around certain constructs:

- `comptime let foo = ...;` on local or global definitions
- `comptime if a { b } else { c }`
- `comptime for ... { ... }`
- `comptime { ... }`

This PR just adds these to parsing and forwards the new node types to
the frontend as necessary.

## Additional Context

The interpreter currently _does not_ handle these properly - this will
come in a separate PR. Eventually we want the interpreter to _only_
interpret comptime things where currently it interprets everything.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.