chore: update NoirJS tutorial (#4861)
# Description

## Problem\*

Resolves #4793

## Summary\*

This pull requests updates the docs to support changes made after
version v0.25.0. The current docs directs the user to install packages
of version 0.19.7 while calling functions `generateProof` and
`verifyProof` instead of `generateFinalProof` and `verifyFinalProof`
which results in the error that functions `generateProof` and
`verifyProof` don't exist.

Also, for the dev version of the docs, using the packages
v0.27.0@nightly, the functions needed to be updated to reflect the
changes in #4769.

## Additional Context

I've tested the changes using packages and nargo of the following
versions:
 - v0.25.0
 - v0.26.0
 - v0.27.0@nightly
 
I added documentation for the `vite.config.js` file otherwise any
version >= v0.21.0 would result in runtime errors.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.