chore: remove pub wildcard import of ast into `noirc_frontend` root (#4862)
# Description

## Problem\*

Related to #4852 

## Summary\*

This PR removes the wildcard import of the `ast` module into the root of
the `noirc_frontend` crate. This encourages us to be more qualified
usage of the ast to make it clearer about which we're working with.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.