chore(experimental): Improve variable not defined error message in comptime interpreter (#4889)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Slightly improves the error message for when a variable in the
interpreter isn't found.

Since all ids are already given by name resolution, the only time one
cannot be found is if it is not a comptime variable, so I've narrowed
the error message a bit.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.