fix: Reset the noir-gates-diff report on master (#4878)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Opened this PR to test
https://github.com/TomAFrench/noir-gates-diff/pull/3 after seeing this
https://github.com/noir-lang/noir/pull/4848#issuecomment-2070241529

## Summary\*

The noir-gates-diff report became malformed on master as "main" was used
as the program name for each program in `execution_success`. This PR
references the last commit in this PR
https://github.com/TomAFrench/noir-gates-diff/pull/3. That PR changes to
using the correct name for each program in `execution_success` but also
resets the report on `master` by comparing against a harcoded reference.
That is why it is expected to see a massive gate diff report on this PR.
The follow-up will revert the changes to accurately check against the
reference.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.