chore(ci): prevent haywire logs (#5966)
Note: code in json sorta sucks, I couldn't comment why this is used.
There is a bad interaction between earthly x github actions x jest
summary reporter. As far as I can see, it's this combination and not any
piece alone (something something ansi codes, but not entirely sure).
This seems like simplest fix.

Without this, it was causing a lot of duplicated lines. But this
particularly interacted with the poor rendering performance of github
actions logs.

The other piece is that its rendering is just slow in chrome for
10,000K+ lines. use raw logs, or another browser like safari