feat: add variable size sha256 (#4920)
# Description

## Problem\*

Resolves #4909 

## Summary\*
Thanks to the sha256compression opcode, we can create conditionally the
blocks to be hashed. As a result there is no additional cost to support
variable input size for sha256, except of course that we pay for the
full input len (even if we hash less).


## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [X] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.