fix: discard ref counts during unrolling (#4923)
# Description

## Problem\*

Resolves #4709 

## Summary\*
The issue is due to reference count instructions which blow-up during
unrolling. Since these instructions are only for Brillig and since loops
are not unrolled in Brillig, I simply discard them during unrolling.


## Additional Context
I did not get an OOM on the mainframe, but the compilation lasted many
hours and did not complete (I finally cancelled it). With the fix the
execution is instantaneous, whether the code is constrained or
unconstrained.


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: jfecher <jake@aztecprotocol.com>