feat(experimental): Implement comptime mut globals (#4926)
# Description

## Problem\*

Resolves #4917 

## Summary\*

Implements mutable comptime globals

## Additional Context

Changes should be relatively straightforward - mostly parser changes to
add parsing `mut` to globals, resolution changes to track this and error
if the global is not also comptime, etc. The one bug that was fixed is
that globals were being re-evaluated in the interpreter each time which
would reset their value whenever they were referenced. This has been
fixed with a lookup check.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>