chore: Add test for recursing a foldable function (#4948)
# Description

## Problem\*

Resolves #4616 

## Summary\*

I wrote the initial issue for recursive calls a while ago. I thought I
had an error at the time and that is why I wrote the issue but it looks
to all be working for fibonacci example. This simply adds a test to
validate recursive ACIR calls.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.