feat: handle empty response foreign calls without an external resolver (#4959)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

I've modified the `DefaultForeignCallExecutor` so that rather than
panicking on an unrecognised foreign call we return an empty response.

This solves an issue experienced by @spalladino where `nargo test` was
failing due to the use of a custom debugging oracle which doesn't return
any values to the circuit. The only situation where these oracle calls
not being handled would change the outcome of the test would be if they
modify state in an external resolver such as to modify the response of a
different oracle call (and so we'd still fail on those anyway).

## Additional Context

I've cleared up some empty `Prover.toml` files as well.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>