chore: Switch Noir JS to use execute program instead of circuit (#4965)
# Description

## Problem\*

Resolves #4706 

## Summary\*

This PR actually does not do too much as we still use
`executeCircuitWithBlackBoxSolver` throughout the ACIR simulator on
aztec. That should be removed separately before we look to remove it
from ACVM JS. Noir JS backend also needs to wait for the barretenberg
interface updates before we can pass a collection of circuits to the
backend.

## Additional Context


## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.