fix: set index and value to 0 for array_get with predicate (#4971)
# Description

## Problem\*

Resolves issue on noir-protocol-circuit and should also resolve issues
described in https://github.com/noir-lang/noir/pull/4716.
The issue it resolves is that array_get under a predicate are done at
index 0 if the predicate is false, but since arrays are not homogenous,
it may contain some value that overflow the type of the array_get.

## Summary\*
When getting value from an array_get, I multiply it with the predicate
to avoid any possible overflow if element at index 0 has not the same
type as the expected one.

## Additional Context
If the array is simple (not nested), I get the offset to a compatible
type and use it when computing the predicate_index.
If the first element of the array has compatible size, then the
predicate_index will be correct
If not, we fallback to the multiplication of the value with the
predicate.


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>