feat: remove query to backend to get expression width (#4975)
# Description

## Problem\*

Progress on #4961 

## Summary\*

This PR removes the query to the backend in order to get the expression
width, for now we just have a blanket default of ~4~ 3 (Looks like we're
still waiting to update to a version of `bb` which uses 4). I've decided
to split the ability to specify a package-level override of this default
as I'd need to do some mucky stuff related to `compile_options`.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.
We don't have docs related to the expression width afaik.
# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.