chore: add script to print lines of code (#4991)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds a small script to print out the LOC for various sections of
the codebase. Adding weightings will come in a later PR. Note that this
PR relies on my fork of tokei which has support for Noir source code.

cc @rumata888

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.